2_coupang_functional/requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
Appium-Python-Client==5.1.1
pytest==8.4.0
requests==2.32.3
selenium==4.33.0